Follow usual type pattern
